Simplifying test examples
