+++ app/build.gradle
@@ -97,7 +97,7 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.6.2'
 
     //Deque AxeDevtoolsAndroid library
-    implementation 'com.deque.android:axe-devtools-android:5.2.0'
+    implementation 'com.deque.android:axe-devtools-android:5.4.0'
 
     implementation 'androidx.activity:activity-compose:1.8.0'
     implementation platform('androidx.compose:compose-bom:2023.03.00')
